added add registry form submit
