+++ deepfence_frontend/apps/dashboard/src/components/registries-connector/DockerConnectorForm.tsx
@@ -1,9 +1,64 @@
+import { forwardRef } from 'react';
 import { HiViewGridAdd } from 'react-icons/hi';
+import { ActionFunctionArgs, Form, redirect, useActionData } from 'react-router-dom';
 import { Card, Step, Stepper, TextInput, Typography } from 'ui-components';
 
-export const DockerConnectorForm = () => {
+import { getRegistriesApiClient } from '@/api/api';
+import { ApiDocsBadRequestResponse } from '@/api/generated';
+import { ApiError, makeRequest } from '@/utils/api';
+
+type ActionReturnType = {
+  message?: string;
+};
+
+export const dockerRegistryConnectorApi = async ({
+  request,
+}: ActionFunctionArgs): Promise<ActionReturnType> => {
+  const formData = await request.formData();
+  const body = Object.fromEntries(formData);
+
+  const r = await makeRequest({
+    apiFunction: getRegistriesApiClient().addRegistry,
+    apiArgs: [
+      {
+        modelRegistryAddReq: {
+          name: body.registryName as string,
+          non_secret: {
+            docker_hub_namespace: body.namespace,
+            docker_hub_username: body.username,
+          },
+          secret: {
+            docker_hub_password: body.password,
+          },
+          registry_type: 'docker_hub',
+        },
+      },
+    ],
+    errorHandler: async (r) => {
+      const error = new ApiError<ActionReturnType>({});
+      if (r.status === 400) {
+        const modelResponse: ApiDocsBadRequestResponse = await r.json();
+        return error.set({
+          message: modelResponse.message ?? '',
+        });
+      }
+    },
+  });
+
+  if (ApiError.isApiError(r)) {
+    return r.value();
+  }
+
+  throw redirect('/onboard/connectors/my-connectors');
+};
+
+export const DockerConnectorForm = forwardRef<HTMLFormElement>((_, ref) => {
+  const actionData = useActionData() as {
+    message?: string;
+  };
+
   return (
-    <>
+    <Form method="post" ref={ref}>
       <Stepper>
         <Step indicator={<HiViewGridAdd />} title="Container Registry">
           <div className={`${Typography.size.sm} dark:text-gray-200`}>
@@ -56,7 +111,9 @@ export const DockerConnectorForm = () => {
           </Card>
         </Step>
       </Stepper>
-      {/* {errorMessage && <p className="text-red-500 text-sm ml-14">{errorMessage}</p>} */}
-    </>
+      {actionData?.message && (
+        <p className="text-red-500 text-sm ml-14 mb-6">{actionData.message}</p>
+      )}
+    </Form>
   );
-};
+});
