+++ deepfence_frontend/apps/dashboard/src/features/onboard/pages/DockerRegistryConnector.tsx
@@ -1,60 +1,17 @@
-import { ActionFunctionArgs, Form, redirect, useActionData } from 'react-router-dom';
+import { useRef } from 'react';
+import { Form } from 'react-router-dom';
 import { Button } from 'ui-components';
 
-import { getRegistriesApiClient } from '@/api/api';
-import { ApiDocsBadRequestResponse } from '@/api/generated';
-import { DockerConnectorForm as DockerRegistryConnectorForm } from '@/components/registries-connector/DockerConnectorForm';
+import {
+  DockerConnectorForm as DockerRegistryConnectorForm,
+  dockerRegistryConnectorApi,
+} from '@/components/registries-connector/DockerConnectorForm';
 import { ConnectorHeader } from '@/features/onboard/components/ConnectorHeader';
-import { ApiError, makeRequest } from '@/utils/api';
 import { usePageNavigation } from '@/utils/usePageNavigation';
-
-type ActionReturnType = {
-  message?: string;
-};
-
-export const action = async ({
-  request,
-}: ActionFunctionArgs): Promise<ActionReturnType> => {
-  const formData = await request.formData();
-  const body = Object.fromEntries(formData);
-
-  const r = await makeRequest({
-    apiFunction: getRegistriesApiClient().addRegistry,
-    apiArgs: [
-      {
-        modelRegistryAddReq: {
-          name: body.registryName as string,
-          non_secret: {
-            docker_hub_namespace: body.namespace,
-            docker_hub_username: body.username,
-          },
-          secret: {
-            docker_hub_password: body.password,
-          },
-          registry_type: 'docker_hub',
-        },
-      },
-    ],
-    errorHandler: async (r) => {
-      const error = new ApiError<ActionReturnType>({});
-      if (r.status === 400) {
-        const modelResponse: ApiDocsBadRequestResponse = await r.json();
-        return error.set({
-          message: modelResponse.message ?? '',
-        });
-      }
-    },
-  });
-
-  if (ApiError.isApiError(r)) {
-    return r.value();
-  }
-
-  throw redirect('/onboard/connectors/my-connectors');
-};
 
 const DockerRegistryContainer = () => {
   const { goBack, navigate } = usePageNavigation();
+  const formRef = useRef<HTMLFormElement>(null);
 
   return (
     <div className="w-full">
@@ -62,8 +19,8 @@ const DockerRegistryContainer = () => {
         title="Docker Registry"
         description="Deploy all modules for Deepfence Compliance Scanner for your registry"
       />
-      <Form method="post">
-        <DockerRegistryConnectorForm />
+      <>
+        <DockerRegistryConnectorForm ref={formRef} />
         <div className="flex">
           <Button onClick={goBack} size="xs">
             Go Back
@@ -74,7 +31,9 @@ const DockerRegistryContainer = () => {
               size="xs"
               className="ml-auto"
               type="submit"
-              onClick={() => {
+              onClick={(e) => {
+                e.preventDefault();
+                formRef.current?.requestSubmit();
                 navigate('/onboard/connectors/my-connectors');
               }}
             >
@@ -82,12 +41,12 @@ const DockerRegistryContainer = () => {
             </Button>
           </div>
         </div>
-      </Form>
+      </>
     </div>
   );
 };
 
 export const module = {
-  action,
+  action: dockerRegistryConnectorApi,
   element: <DockerRegistryContainer />,
 };
